refacotor: move mock to network from domain
