+++ core/data/build.gradle.kts
@@ -19,14 +19,6 @@ android {
             dimension = "env"
         }
     }
-    sourceSets {
-        getByName("prod") {
-            java.srcDirs("src/prod/java")
-        }
-        getByName("mock") {
-            java.srcDirs("src/mock/java")
-        }
-    }
     ksp {
         arg("room.schemaLocation", "$projectDir/schemas")
     }
