+++ core/network/src/prod/java/ksnd/hiraganaconverter/core/netrwork/impl/ConvertApiClientImpl.kt
@@ -1,4 +1,4 @@
-package ksnd.hiraganaconverter.core.network.impl
+package ksnd.hiraganaconverter.core.netrwork.impl
 
 import com.jakewharton.retrofit2.converter.kotlinx.serialization.asConverterFactory
 import kotlinx.serialization.encodeToString
